MAYA-106486 Handle UV linkage on material export
We now explore the UV linkage and make sure the UV stream name in the material matches the UV set name on the shape.
We will create as many material specializations as needed to match the shape streams in case of UV stream names inconsistency in a group of shapes sharing a single material.
We no longer automatically rename "map1" to "st" since the UV streams in the material will now be correctly set.