cache empty values for unused physical camera attributes in Pixar's batch renderer
HdCamera::Sync() now pulls on all of these physical attributes, so the Pixar
batch renderer provides empty values for everything that is unused. The
Viewport 2.0 render delegate does not use this interface.

The "exposure" camera token was added in core USD 20.11 and the "projection"
camera token was added for core USD 21.02.

This change corresponds to core USD commit
https://github.com/PixarAnimationStudios/USD/commit/84cacd09d1a0e0a1886b2a2494a16ad2625898ba

(Internal changes: 2124799, 2124988)