ensure proper handling of point instances in UsdHierarchy
When a scene item represents a point instance, its Ufe::Path will end with a
path component that is all digits representing the instance index (e.g.
".../PointInstancer/12345"). Point instances are generated by a PointInstancer
prim based on data members on the PointInstancer and the prototype being
instanced. There is otherwise no direct mapping of a point instance to any
location in USD namespace.

As a result, we treat point instances as child-less in UFE. For the same reason,
point instances don't really have a "parent" in USD either, but in UFE we
consider the PointInstancer prim the "parent" of a point instance it generated.
This allows pick-walking from a point instance up to the PointInstancer prim
that generated it. Note that the reverse is *not* true. The point instances
generated by a PointInstancer cannot be retrieved by querying the
PointInstancer for its children.