support install/remove of UsdMayaDiagnosticDelegate across multiple plugins
This adds an installation count that is used to determine whether to construct
or destroy the shared UsdMayaDiagnosticDelegate when multiple calls to the
static InstallDelegate() and RemoveDelegate() functions are made. This ensures
that only a single instance of the delegate is installed when for example both
the "pxrUsd" and "mayaUsdPlugin" Maya plugins are loaded, and that the delegate
is not removed until both plugins are unloaded.

The installation counter follows a similar pattern to the one used for
MayaUsdProxyShapePlugin.