replace #!/pxrpythonsubst with #!/usr/bin/env mayapy
Now that the shebang.py script is gone, there's no clue as to the meaning
pxrpythonsubst.  The project never makes use of the shebang line anyway,
since we support windows, so we always explcitly call mayapy with the
script as an arg... so we could eliminate it altogether.  But by leaving
it as mayapy, it may at least be occasionally useful for testers running
things manually... and at least documents that it should be run with
mayapy.