migrate testUsdMayaProxyShape from pxr to mayaUsd as testProxyShapeBase
This test is applicable to all flavors of mayaUsdProxyShapeBase. Though it is
by no means an exhaustive test in its current form, it provides a base that can
be extended with additional tests for the base proxy shape that is common to
all derived proxy shapes. Tests for functionality specific to derived proxy
shapes should go in separate tests.