MAYA-108707 Share shader effect among similar-enough material networks
Replace the authored node paths with simplified paths in the form of "node#".
By doing so we will be able to reuse shader effects among material networks
which have the same node identifiers and relationships but different node
paths, reduce shader compilation overhead and enable material consolidation
for faster rendering.