fix regex in testDiagnosticDelegate to handle varying ways of executing the test
This is a follow-up fix to a change I made in #1042 that replaced the
"__main__" module name included as part of a coding error message with
"testDiagnosticDelegate" instead. While this is correct for the way the test is
run using the mayaUsd CMake testing infrastructure, it is not correct in all
cases.

When the test is executed as a result of the unittest.main() call at the bottom
of the file (e.g. when the script is executed directly via its shebang, when
the script file is passed to a Python interpreter on the command-line, or when
the script file's contents is read and executed through compile() and exec(),
the module will be reported as "__main__". Otherwise, when the file is imported
as a module and then executed using unittest.main(module=<module name>), the
module name will match the file name and be reported as
"testDiagnosticDelegate".

To preserve the extra detail in the test, the regex used to match against the
exception message was updated to recognize both cases.