set kind=group on the UsdPrim created by the group command when its parent is in the model hierarchy
Authoring the kind metadata on the group Xform prim ensures that the model
hierarchy remains contiguous. Prims underneath the group prim may still be in
the model hierarchy if they also have kind authored.

If the parent prim of the group prim is not part of the model hierarchy, no
kind is authored on the group prim.