MAYA-109316 Propagate state of dirty draw items
when they are about to be redrawn after getting out of sync.

This fixes drawing multiple Repr in sequence after dirtying the RPrim.
The first draw will clean the RPrim and the Reprs that were selected
to be drawn. Reprs that were not drawn will stay stale and need to
signal the Sync engine that they need to be refreshed on the next
draw cycle that will draw them.