implement USD point instances pick modes
These changes implement the "PointInstancer", "Instances", and "Prototypes"
point instances pick modes. If we did not pick a point instance after applying
the point instances pick mode, we then apply any USD selection kind filtering
if necessary.

We also ensure that the resolved scene item's path is created appropriately
using the usdPathToUfePathSegment() utility function.

If the point instances pick mode optionVar is unset, the default behavior is
to pick PointInstancers. This differs from the previous behavior which was to
pick prototypes.