make tests in test/lib/ufe directly runnable and use fixturesUtils
This change syncs the tests in test/lib/ufe with tests elsewhere in the source
tree so that the script files are directly runnable (with the addition of a
__name__ == "__main__" block at the bottom) rather than only as a module. They
also now use fixturesUtils to ensure that Maya standalone is initialized and
uninitialized in that case.

This commit also includes some clean-up and reordering of the module imports at
the top of each test script.