MAYA-105623 better support for USD leftHanded mesh.
The  main problem is that the 'leftHanded' orientation is USD changes the interpretation of the vertices order of faces *and* the direction of generated normals.

On the other hand, the 'opposite' flag in Maya flips the faces directions, but does not flip the normals.

So the behaviour differs. To better suport USD import of externally-generated USD files, we will no longer set the 'opposite' flag when importing ledftHanded USD files. We will instead change the vertices orders to be right-handed.

Note that we will still export 'opposite' geometries from Maya as by setting the 'leftHanded' flag. This is done for two reasons: first, to keep the code change minimal, second because USD does not support different handedness for faces and normasl anyway.