re-implement UV set export to maintain UV shells and connectivity
Previously when exporting UV sets to USD, we used the face vertex ordering
imposed by MItMeshFaceVertex to effectively re-arrange all of the Maya mesh's
UVs into face vertex order. We then further de-duplicated the UV values and
compressed the indices array to a coarser interpolation if possible. All of
this manipulation effectively destroyed any UV shell and connectivity
information that was encoded in the Maya indexing, resulting in UV sets that
were disconnected and difficult to work with if a Maya mesh was round-tripped
out to USD and back in.

Instead, we now translate the Maya mesh's UV values and assignment indices more
or less directly, and we no longer apply any compression. This offers the best
chance of maintaining the UV set's connectivity across round-trips. The net
result in USD is the same (i.e. the array of UVs produced by doing
primvar.ComputeFlattened() is the same as before), but the array values for the
UV set primvar and its accompanying indices attribute may be different.