apply workaround in testUsdExportUVSets to ensure that UV value arrays are of minimal length
Although the UV sets on the "SharedFacesCubeShape" are stored in the Maya scene
with a minimal number of UV values and UV shells, they seem to be expanded when
the file is opened such that we end up with a UV value per face vertex rather
than smaller arrays of UV values with only the indices being per face vertex.
As a result, we have to reassign the UVs just before we export.