add an abstract base class for objects that modify attributes of PointInstancers
This adds a UsdPointInstanceModifierBase class which will be specialized to
create modifiers for the "positions", "orientations", and "scales" attributes
of PointInstancer prims. Derived classes of this base class should implement
the functions to directly get and create the USD attribute on the
PointInstancer (typically with a call to UsdGeomPointInstancer::Get...Attr()
and UsdGeomPointInstancer::Create...Attr(), respectively), functions to convert
values between USD and UFE types, and a function to retrieve the default value
in the USD type.