move import of ufeSelectCmd inside selectPath() function in ufeUtils
Attempting to import ufeSelectCmd before maya.standalone.initialize() has been
called will raise an ImportError and cause tests to fail. For example, this
would be the case for test cases that are decorated with the
unittest.skipUnless() decorator and use ufeUtils.ufeFeatureSetVersion(), since
the decorator will be executed before setUpClass has a chance to run and
initialize maya.standalone.

This change addresses that by pushing the import of ufeSelectCmd into the
selectPath() function where it is used rather than having it at the top of the
file. This ensures that the ufeUtils module can be imported and
ufeUtils.ufeFeatureSetVersion() can be called safely before
maya.standalone.initialize().