issue Transform3d notifications for point instance manipulation in response to USD notices
When individual point instances are manipulated, the change is represented in
USD as an update to the "positions", "orientations", or "scales" array
attribute(s) on the PointInstancer prim that generated the point instance.

Unfortunately though, there is no way for us to know which point instance(s)
were actually affected based on the USD change notification. This means that we
have to assume that *all* of the PointInstancer prim's instances may have been
affected. We therefore need to construct UFE paths for every instance and issue
a notification for each one.