MAYA-110917 Increase depth priority for point snapping items
After turning on camera-based selection and zooming out far enough
to a USD object, we could only snap to silhouette points. After
zooming in closer, the issue will become better. This is a preexisting
depth priority issue in VP2RenderDelegate so it has nothing to do with
the recent point snapping refactoring inside Maya (MAYA-109518).