Compatibility with the way how Alembic used to handle the IO of the User Attributes And the Hierarchyduring the I/O operations, the aim of these features is to simplify and reduce the amount of work for the migration from Alembic to USD pipeline.
Features:
---------
* Added -userattr/-u multi-flag to export user custom attributes to usd.
* Added -root/-rt multi-flag that copies the behavior of AbcExport which limits the roots under which the primitives will be written in the USD rather than translating all the parents up till scene root.
* All importer/translators will import all custom attributes found in USD to Maya when loading USDs in anyway (including assembly), the created attributes in Maya on import, will not have any namespace in the attribute name, just like how it was when it was exported at the first place.

Enhancements:
-------------

* -userattr/-u (multi-flag) allows passing the custom user attributes to be exported in USD without having to add a special tags on the exported nodes themselves, the value of the flag should have the desired namespace where the attribute will be written, for example::

usdExport -userattr "userProperties:ast" -userattr "userProperties:context" ....
usdExport -userattr "primvars:varyColor" ....

This will check the maya nodes being exported, if the attr (without the namespace) exists, then it will be exported under the given namespace, if it's not there for a specific object, the flag is ignored.

If the attribute being exported found on a transform node, it will be exported under an 'xform:' namespace, for example with: -userattr "userProperties:ast", if the the maya attribute "ast" found on transform node, it will be written in USD under xform:userProperties:ast, if it's found on a shape node, it will be written as is to userProperties:ast (and of course if doesn't exist, will be ignored).

This way even when merging transform/shape on the exported USD, we can still recognize if the 'ast' was a transform attribute or a 'shape' attribute especially when we import these attributes back again (which is another feature found in this same branch).

Note: when using the -userattr method, the maya doubles precision goes to USD as is without conversion, there is no option in this way to set additional information about the exported attrs as the case with normal attribute exporting using the json method, however, most of the attributes especially the type is set automatically based what we found in Maya.

* all pixar, maya, and Animal logic USD import/translators will create all the Maya custom attributes found in the loaded USD (in case with the attributes found under userProperties: namespace, it will all be created on the shape node, if it was under xform:userProperties: namespace, the it will be created in Maya on the transform node of the object.

* -root/-rt (multi-flag) and -sl flag together gives full control on which part of the hierarchy to be written to the USD and what exactly under the given roots will be exported, for example:
if we have the following hierarchy:

Assets
|---- chr_model_ast
|----|---- body
|----|---- aux_nodes
|----chr_garment_ast
|----|---- cloth
|----|---- accessories
|----|---- garment_curves_grp
|----|---- sim_garment

select -r body cloth accessories;
usdExport -root "chr_model_ast" -root "chr_garment_ast" -userattr "userProperties:ast" -userattr "userProperties:version" -sl -file "/path/to/usd.usda";

this will write a USD that have the following hierarchy (with the custom attributes exported):

/
chr_model_ast
|--- body
chr_garment_ast
|--- cloth
|--- accessories