MAYA-112059 - UsdUVTexture works with Maya color management prefs
Maya versions before 2022 had a default color management rendering space of "scene-linear Rec 709/sRGB", which means converting a sRGB texture required only the gamma correction part,
With Maya 2022, we now use the "ACEScg" rendering color space, so an extra multiplication by a color correction matrix is required to move from linrec709 to ACEScg.
This is a user preference, though, and one that is saved with the scene. So we need to query the current rendering space and use the correct color correction.

This is the list of supported rendering color spaces:
- scene-linear Rec 709/sRGB
- ACEScg
- ACES2065-1
- scene-linear DCI-P3 D65
- scene-linear Rec.2020

Note that the VP2 render delegate does not register for notifications when the rendering color space is changed, so if you change the rendering color space, you should save the scene and reload to force the USD stages materials to be updated with the correct color space transformation.