MAYA-112227 - UV set renamed to st, st1, st2
Implements automatic UV set renaming.
- The sets will be renamed st, st1, st2, st3,... in the order they are found in the mesh.
- The original name will be stored in custom data for roundtripping
- On import, the names will always be restored if the roundtripping information is found
- On import, the UV sets are restored in the same order as found in the USD primitive
- The environment variables MAYAUSD_IMPORT_PRIMARY_UV_SET_AS_MAP1 and  MAYAUSD_EXPORT_MAP1_AS_PRIMARY_UV_SET were removed as their functionnality is subsumed in the global renaming.

Caveat:
If you have a mesh with multiple UV sets, and a UV set named "map1" is not the first UV set, then it will not automatically be renamed "st" as was previously the case using MAYAUSD_EXPORT_MAP1_AS_PRIMARY_UV_SET. We expect these cases to be quite rare, but if you have post-processing scripts that are expecting "st" and scenes where "map1" is not the first UV sets; then the results at the other end of the post-processing pipeline will be different.