MAYA-112769 add prim selectability temporary cache and change selectability state names
The cache is cleared at the beginning of each selection to avoid the always thorny problem of cache invalidation.
The cache accelerate the verification of selectability in bug scene of thousands of prims being all region-selected.

Change "selectable" to "on", Change "unselectable" to "off".
This makes the state more generic in case we ant to add other inheritable metadata. They will be able to follow the same naming convention: metadata-name = "on" | off" | "inherit"