MAYA-113559 prevent crash when accessing a stale attribute from Python
User (and ours) Python scripts can keep variable holding references to invalid USD attributes. Detect when an invalid USD attribute is accessed and prevent getting and setting its value rather than using invalid pointers and crash.

This was triggered in our own UI code when receiving notifications of attribute value change when two scenes with prim with identical UFE paths were manipulated in successsion in a Python script.

A typical example was a Python script that created a new Maya scene, a new USD stage, and a cone within that stage and setting a value on a property. Re-running the script within the same Maya session would crash.