Fixing lighting issues with UsdPreviewSurface
We recently debugged a scene that failed to render correctly with the V2
Light API. The main issue with shadows was fixed in another PR, but there
were still 3 other problems where the Light API V1 was not on-par with USD
and the V2 Light API.

1- Directional light affects IBL lighting

Create a test scene:

- 20x20 plane at the origin, usdPreviewSurface shading, 50% metalness
- aiSkyDomeLight with an environment map
- Directional light with *zero* intensity

Rotate the directional light, and watch how it affects the lighting of the plane. That light has zero intensity, so it should have no effect.

2- textured occlusion does not affect environment lighting

One of the assets in the scene had a textured occlusion map which
correctly worked with scene lights, but not with aiSkyDome lights.

3- specular highlights from regular Maya lights did not have enough base
color for metallic workflows.

That one was fixed for aiSkyDome lights in PR #1356 but direct lighting
still used the old formula.