MAYA-109415 Fix warning on blendShape export
Got "Unrecognized node" error on blendShape export. The node was the
fCurve driving the weight.

Looked in mayaFindOrigMeshFromBlendShapeTarget() and found out that the
traversal would start on a random plug in the checked branch. Sometimes
the input geom would be found first, sometimes the fCurve would be
traversed as well.

Fixed the code to make sure the traversal starts at the inputGeom plug
that would have been used to compute fnGeoFilter.inputShapeAtIndex(...)
in the ignoreWarning branch.