MAYA-112770 return null cmd to convery the attribute is locked
Like other UFE commands-getters, use the fact that the setCmd() returns null to convey to the caller that the attribute is not editable.

Adjust the corresponding test.

One unfortunate side-effect is that every call to setCmd() must be checked for nullptr (in C++) or None (in Python), making the calling code more verbose.