MAYA-114111 Added relationship comparison
Currently, the relationship comparison algorithm is simple. Given the
generality of list-op data, we would need a complex algorithm to compute
minimal diff. Instead, we assume that most case will be simple additions at
the front or end of an existing list, or removing some items. We also assume
most lists will be short, so it is not worth it to write a generic complex
algorithm. The algorithm detects prepend, append and removal.

- Added Prepended and Appended to possible DiffResult.
- Changed DiffResultMap to DiffResultPerToken.
- Added DiffResultPerPath, for relationship comparison result.
- Added DiffResultPerPathPerToken, for multiple reltionships results.
- Added comparePrimsRelationships().
- Added compareRelationships() to compre two relationships.
- Added computeOverallResult() to combine multiple sub-results into an overall result.
- Added unit tests for ovall results.
- Added unit tests for prims relationships comparisons.
- Added unit tests for relationships comparisons.