Use HdRprim's render tag data for VP2 mesh shared data
Starting in USD 22.02, Hydra internally manages render tag data for
rprims. This was causing the dirty bit to be cleared prior to calling
sync, causing the render tag data to never get set in _meshSharedData.
Ultimately we could always just use the rprim's GetRenderTag API, but
since MeshViewportCompute is only sent the SharedData strcuture, we'll
set _meshSharedData unconditionally for now.