MAYA-114297 MAYA-114300 Job context UI
Add a job context dropdown called "Plug-in Configuration:" in both
import and export dialogs when in USD mode.

This dropdown will be populated with the available job contexts.

Selecting on the the job context will populate the UI with the settings
provided by the job context, and these settings will be made read-only.