MAYA-114995 avoid merging equivalent transforms.
Compare the full local transform when deciding to merge transform-related properties to avoid authoring transforms when they did not change but their representation changed.

Get the union of all time codes of all transform properties and compre the value at all time codes.

Add a unit test. Confirmed that the test fails when the new code is disabled.