MAYA-114109 fix merge-to-USD
When targeting the session layer and a non-root prim, the hierarchy above the prim no longer existing by the time the data was being merged. Move the removal of the pull info to after the merge so that the hierarchy still exists in the session layer during the merge.

Add a unit test that filed without this fix.