MAYA-114898 fix unit test assertions related to pull/push
When using the direct python API to edit-as-Maya and merge-to-USD, the calls have to be wrapped in a OpUndoItemList block to capture the undo info, otherwise it will linger on. This is not a problem, except when Maya exit, the order of cleanup makes Maya assert that some deleted nodes are still present.

Fix the tests by using a block.

Also do a best effort to clean up the undo information that may have been left behind due to not using a OpUndoInfoList block in Python. On scene reset the undo info is cleared, although currently the assertions are hit before the scene-reset callback is triggered, so we still get assertions.