MAYA-121242 fix flipped normals.
Maya did not write the USD "orientation" attribute. In addition, Maya flipped the vertices order when editing a USD mesh with "leftHanded" orientation. On merge-to-USD, Maya would write right-handed vertices but not write the "orientation" property.

The result would be righ-handed verticeswith left-handed orientation attribute.

This makes multiple fixes: instead of flipped vertices order, we now set the "opposite" flag on import. This is what the nurbs were already doing. More testing will be required to see what effects, if any, this may have on normals.

The second fix is to explicitly write the "orientation" attribute on export.

A few improvements in the verbosity options and simplification of some code were also made whilte investigating the problem and are included.