MAYA-121242 fix flipped normals.
Maya did not write the USD "orientation" attribute. In addition, Maya flipped the vertices order when editing a USD mesh with "leftHanded" orientation. On merge-to-USD, Maya would write right-handed vertices but not write the "orientation" property.

The result would be righ-handed verticeswith left-handed orientation attribute.

See: MAYA-105623 for the original reason we flip vertices orders. It has to do with the way "leftHanded" affect vertices order and normals in USD.

The fix is to explicitly write the "orientation" attribute on export.

A few improvements in the verbosity options and simplification of some code were also made whilte investigating the problem and are included.