Fix import of UsdLuxLight Prims
Fixes a bug where UsdLuxLight prims underneath a Scope
were not being imported properly (if the Scope only had light prims
underneath it).

The import logic for "Scope" uses a heuristic to see if that scope is a
"Looks" or "Materials" scope.  It did this by checking if any of the
children nodes were UsdShadeConnectableAPI.  Unfortunately, lights (and
other non-shading prims) can be UsdShadeConnectableAPI.

This new heursistic checks for Material prims.  Note, this was more or
less the original version that Autodesk proposed in:
https://github.com/Autodesk/maya-usd/pull/985