MAYA-151514 Add before-reset notification.
The after-scene-reset is necessary for the render code but the before-scene-reset is necessary for the undo items. So different code need each one, that is why we have both now.

Replacing the AfterNew with BeforeNew makes some tests fail on Linux. For example testProxyShapeRendererSceneMessages. They crash deep into OGS, in its callback that reach mayaUsd imaging delegates.