Fixed Bind contextual menu slowdown
- Traversing large scenes is not an option. Instead we traverse the
  selection to find materials either directly selected, or under a
  selected parent (Looks section).

The workflow is now:
 - Select the materials or the looks section
 - Navigate in the outliner to the geometry you want to bind, and
   activate the context menu (this does not change selection).
 - If materials are found in the selection, they will be offered in a
   Bind Material context menu item.

In UX terms, this forces you to pre-select the "object" before right
clicking on the "subject" to get the Bind "verb" as an option. This is far
from perfect, so I am open to any other solution that does not process the
whole stage to find materials.