MAYA-122541 fix hydra crash with skydome
Some resources used by the Hydra renderer are cached in the Hydra engine task context. The shutdown of the Maya-to-Hydra render override destroyed many Hydra-related objects, but the Hydra engine itself is kept by value and is never destroyed. We ended-up having data in teh task context that still referred to deleted objects and that led to a crsh if the Hydra render override was initialized and used again.

Arguably, the bug is in Hydra itself, where the texture handles (HdStTextureHandle) keep a raw pointer to the texture registry (HdSt_TextureHandleRegistry) instead of a smart pointer. That raw pointer is taken from the HdRenderIndex, which does keep its pointer to the registry as a shared_ptr.

OTOH, semi-shutting-down Hydra like our render override does is probably not correct. These low-level interactions and handling of shutdown and restart are rarely well documented.

In passing, I also fixed a bug in the UFE helper that assumed there always was at least one segment in teh given UFE path, which tured out to be wrong in some cases.