MAYA-115146 avoid problem with clashing Maya Ref
Improve the code to avoid having two Maya Ref that happen to refer to the same reference file use each other's Maya Ref Node by accident.

The subtle change is that we no longer search for the corresponding Maya Ref Node if the custom attribute is not present. When the attribute is not present, it means the Maya Ref was never edited and thus need to create its own Maya Ref Node, not search for an existing one.

I also refactored the code to avoid some code duplication and push out some code into their own functions for clarity.

- Make all ref-node-name generator always generate the full name, including any RN suffix.
- Renamed isMayaReference() to useNewMayaRefNaming() for clarity since that was the function's purpose.
- Added setMayaRefCustomAttribute() and getMayaRefCustomAttribute() for clarity, isolation and reuse.
- Got rid of getRefNodeNameFromPrim() since its only usage was to try to find a Maya Ref Node when the custom attribute was absent, which was incorrect and led to accidentally use the node from other Maya Ref.
- De-duplicate the Maya ref node searching code.
- Fixed the new test that simulated reparent a prim but did not simulate that the custom attribute would follow.