MAYA-123016 stage clear now resets the outliner
When a prim is shown in the outliner and  the outliner gets collapsed, then the outliner still keeps the reference to the prim. But because it is collapsed it does not update when the selection changes. On the other hand, stages only go away when a new Maya scene is created since they are cached. Also, UFE notifications are based on UFE paths.

So if a USD file was staged, a prim selected and shown in the AE, then the AE gets collapsed, then a new Maya scene gets created and the same USD file is staged, then selecting the same prim and modifying (say, translating it), it would send a UFE notification with the same UFE path, but the prim in the AE is the prim in the old stage that no longer exists.

We now reset the outliner when clearing the stage cache for this reason.