MAYA-123419 add execution validation during pull/push
The undoable execution items were not returning if the execution was successful from the convenience static execute function, preventing the validation of correct execution.

Fix this by making the execute function of PythonUndoItem and FunctionUndoItem return a boolean success flag, received frmo the redo function. Use this flag during pull and pull to validate that key steps executed correctly.

In our case, the authoring of metadat on an instance prototype is forbidden by USD and was failing, but that fact was lost. So no pull information was written, which entirely broke the pull/push workflow.