MAYA-123509 fix merging of normals
Normals in USD can be kept in the "normals" attribute or the "primvars:normals" attribute. This breaks the semantic-agnostic merge algorithm as it did not relate the two attributes.

In addition, the "privars:normals" has higher priority.

The result was that the round-trip of Edit-as-Maya and Merge-to-USD would create the "normals" attribute attribute and the resulting prim would have both attributes, but the old unmodified primvats would win.

This was bad when the mesh was edited as the wrong normals were used, but catastrophic when the topology was edited as the number of normals was incorrect.

The fix is to add custom logic to handle normals merging. The new strategy is to rely on the USD exportdone during merge to always author the "normals" attribute and to delete the existing "primvars:normals" from the original USD data so that it does not take priority on the newly edited normals.

While fixing this, other small fixes were done:
- the messages printed during merging were improved to be easier to understand.
- Added support for some additional data-type during USD import: byte and unsigned int were not supported.
- Added a comments about the fact that we currently do no compare metadata to determine if a prim has changed.