MAYA-123579 fix crash when deleting a stage
The loop that retrieved all the stage was calling a function that could modify the container that was iterated over. Fix this by first caching the items we want to iterate.

It only crashed in debug build due to the luck that the meory was still present and accessible in release build.

Added a unit test that crashed in debug and now works with the fix.