MAYA-114376 preserve unsaved root layer
When the proxy stage is recompute with an unsaved root layer, the attribute keeping the root layer name would be empty. A new root layer would be created each time, losing any previous content.

This attribute only gets set once the stage gets saved.

Fix this by re-using the same anonymous root layer when the stage has not been saved yet.

Also, as a bonus, this fix change teh code so that the same code path for newly created stage is used, which means that if the user ahas set the option var to target the session layer, it will also be honored for newly-created stages. It will also re-use any pre-existing stage, so we also avoid creating stages repeatedly for the same unsaved proxy stage node every time an attribute get sets.