MAYA-123815 prevent usage of invalid prim or sibling range.
When the USD prim is deactivated, many operations will crash if called on that prim, as the USD prim API assumes that many of its functions are illegal to call on an invalid prim abd deactivated prim are invalid.

(Internally, the UsdPrim contains a pointer which is null when invalid, and most function don't check the pointer, instead relying on the caller to not call them when invalid.)

So check the prim validity in parent() and when creating sibling iterators.