MAYA-123792 fix select undeleted prim
When a stage was delete in the outliner and then the deletion undone, selection of prim would cause errors. Also, the outliner content would be incorrect.

The root cause was that undo would not update the stage cache, which would break a lot of thing.

The fix is to not try to catch every possible ways the stage might get invalid (the famous "cache invalidation is one of the hard problem of computer science") and instead just invalidate and refilled it when a cache miss happens.

As the stage cache is not a LRU cache but a mapping between representations, its content is really rarely incorrect, except during transient period during the flurry of notification affecting its validity: deleting a stage node, renaming it, undoing these operation, etc. These operation would ebd-up invalidating the cache anyway, so rebuilding it on miss is equivalent.

Each approach has its pro and con. Using explicit invalidation theoretically minimizes the number of rebuilds, but require to perfectly catch every situation that may make the cache invalid, which is error prone and costly to design and maintain. Using automatic rebuilding could theoritically cause mutiple rebuilds, but does not require any code maintenance.

The "theoretically" are important there. In actual usage, automatic rebuild are efficient, because the transient period is short and in my testing the cache gets rebuilt only once. (i.e. some sub-system tries to access a stage that was undeleted, the cache is rebuilt and then the cache works for all subsequent callers.) Ironically, trying to explicitly invalidate the cache leads to multiple invalidations because the notifications sometimes come from multiple sources, so are triggered multiple times.